change header file

+++ typewriter/Editor.h
@@ -48,7 +48,7 @@
         void move_right();
         void move_left();
         void insert(int c, int line, int col);
-        void delete_char(int line, int col);
+        void delete_char(size_t line, size_t col);
         void command_mode();
         void backspace();
         void new_line();
